wip: phase 5 — extract useAuthGuard hook, replace 4 inconsistent auth patterns
